Split cast tests into CastTest class
