+++ .travis.yml
@@ -15,4 +15,3 @@ script:
 after_success:
   - make test
   - make run
-
